+++ web/libs/editor/src/regions/Area.js
@@ -15,7 +15,7 @@ import { TimelineRegionModel } from "./TimelineRegion";
 import { TimeSeriesRegionModel } from "./TimeSeriesRegion";
 import { ParagraphsRegionModel } from "./ParagraphsRegion";
 import { VideoRectangleRegionModel } from "./VideoRectangleRegion";
-import { PixelWiseRegionModel } from "./PixelWiseRegion";
+import { PixelWiseRegionModel } from "./PixelWiseRegion/PixelWiseRegion";
 
 // general Area type for classification Results which doesn't belong to any real Area
 const ClassificationArea = types.compose(
