+++ web/package.json
@@ -82,13 +82,16 @@
     "lodash.get": "^4.4.0",
     "lodash.ismatch": "^4.4.0",
     "lodash.throttle": "^4.1.1",
+    "marching-squares": "^1.0.0",
     "mobx": "^5.15.4",
     "mobx-react": "^6",
     "mobx-state-tree": "^3.16.0",
     "nanoid": "^3.3.8",
+    "oslllo-potrace": "^4.0.0",
     "pako": "^2.1.0",
     "papaparse": "^5.4.1",
     "pleasejs": "^0.4.2",
+    "potrace-wasm": "^1.0.4",
     "rc-tree": "^5.7.8",
     "react": "18.2.0",
     "react-beautiful-dnd": "^13.1.1",
@@ -107,6 +110,7 @@
     "react-window-infinite-loader": "^1.0.5",
     "sanitize-html": "^2.14.0",
     "shadcn": "^2.1.8",
+    "simplify-js": "^1.2.4",
     "storybook": "^8.5.0",
     "strman": "^2.0.1",
     "tailwind-merge": "^2.6.0",
